Ported LoRaWAN Stack

+++ HW_BSP/inc/hw_adc.h
@@ -5,6 +5,6 @@ void HW_Adc_Init( void );
 void HW_Adc_DeInit( void );
 
 uint16_t HW_GetTemperatureLevel( void );
-uint16_t HW_GetBatteryLevel( void );
+uint8_t HW_GetBatteryLevel( void );
 
 #endif /* INC_HW_ADC_H_ */
